:shirt: Small refactor, update snippet and upgrade code analyzer
