+++ .vscode/header-cs.code-snippets
@@ -23,7 +23,7 @@
 			"// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,",
 			"// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE",
 			"// SOFTWARE.",
-			"namespace ${1:SceneGate.Games.ProfessorLayton}",
+			"namespace ${1:Texim}",
 			"{",
 			"    $0",
 			"}",
