+++ .vscode/settings.json
@@ -14,6 +14,7 @@
     },
     "prettier.proseWrap": "always",
     "cSpell.words": [
+        "Argb",
         "Texim"
     ],
 }
