+++ src/Directory.Packages.props
@@ -9,7 +9,7 @@
         <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="1.0.0" />
 
         <PackageVersion Include="StyleCop.Analyzers" Version="1.1.118" />
-        <PackageVersion Include="SonarAnalyzer.CSharp" Version="8.19.0.28253" />
+        <PackageVersion Include="SonarAnalyzer.CSharp" Version="8.20.0.28934" />
         <PackageVersion Include="Roslynator.Analyzers" Version="3.1.0" />
     </ItemGroup>
 </Project>
