+++ src/Texim/Pixel.cs
@@ -85,36 +85,4 @@ namespace Texim
             return string.Format("[Pixel: IsIndexed={0}, Info={1}, Alpha={2}]", IsIndexed, Info, Alpha);
         }
     }
-
-    public static class PixelExtension
-    {
-        public static void FlipX(this Pixel[] tile, Size tileSize)
-        {
-            for (int y = 0; y < tileSize.Height; y++) {
-                for (int x = 0; x < tileSize.Width / 2; x++) {
-                    int t1 = y * tileSize.Width + x;
-                    int t2 = y * tileSize.Width + (tileSize.Width - 1 - x);
-
-                    Pixel swap = tile[t1];
-                    tile[t1] = tile[t2];
-                    tile[t2] = swap;
-                }
-            }
-        }
-
-        public static void FlipY(this Pixel[] tile, Size tileSize)
-        {
-            for (int x = 0; x < tileSize.Width; x++) {
-                for (int y = 0; y < tileSize.Height / 2; y++) {
-                    int t1 = x + tileSize.Width * y;
-                    int t2 = x + tileSize.Width * (tileSize.Height - 1 - y);
-
-                    Pixel swap = tile[t1];
-                    tile[t1] = tile[t2];
-                    tile[t2] = swap;
-                }
-            }
-        }
-    }
 }
-
